Config Publish

+++ app/Livewire/Admin/Users/UserIndexComponent.php
@@ -32,8 +32,13 @@ class UserIndexComponent extends Component
     public function ChangeType($id)
     {
         $user = User::findOrFail($id);
-        $user->utype = $user->utype == 'USR' ? 'ADM' : 'USR';
-        $user->update();
+        if (!$user->admin) {
+            $user->utype = $user->utype == 'USR' ? 'ADM' : 'USR';
+            $user->update();
+        } else {
+            session()->flash('error', 'Error.');
+            $this->redirectRoute('admin.users.index', navigate: true);
+        }
     }
 
     public function editUser($id)
@@ -51,8 +56,13 @@ class UserIndexComponent extends Component
     public function destroy()
     {
         $user = User::findOrFail($this->delId);
-        $user->delete();
-        $this->dispatch('closeNewsModal');
-        session()->flash('error', 'Данные пользователя удалены.');
+        if (!$user->admin){
+            $user->delete();
+            $this->dispatch('closeNewsModal');
+            session()->flash('error', 'Данные пользователя удалены.');
+        }else {
+            session()->flash('error', 'Error.');
+            $this->redirectRoute('admin.users.index', navigate: true);
+        }
     }
 }
